requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
sqlparse==0.5.2
typing_extensions==4.12.2
